requirements.txt
Django==5.1.4
djangorestframework==3.16.1
sqlparse==0.5.4
tzdata==2025.3
psycopg[binary]
gunicorn
Pillow>=10.0.0
celery>=5.3
redis>=5.0
